move invoice props to Document
